backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
Werkzeug==3.0.1
Pillow==10.2.0
numpy==1.26.3
tensorflow==2.15.0
python-dotenv==1.0.0
gunicorn==21.2.0
